add: inventory view

+++ cmd/table/application/TableDelete.go
@@ -1,6 +1,7 @@
 package application
 
 import (
+	"restaurant-management-backend/cmd/table/domain"
 	"restaurant-management-backend/cmd/table/domain/repository"
 	"restaurant-management-backend/cmd/table/domain/types"
 )
@@ -18,5 +19,10 @@ func (this TableDelete) Execute(id int) error {
 	if err != nil {
 		return err
 	}
+
+	if table, _ := this.repository.GetById(tableId); table == nil {
+		return domain.NewTableNotFound(*tableId)
+	}
+
 	return this.repository.Delete(tableId)
 }
