+++ cmd/user/application/UserDelete.go
@@ -1,6 +1,7 @@
 package application
 
 import (
+	"restaurant-management-backend/cmd/user/domain"
 	"restaurant-management-backend/cmd/user/domain/repository"
 	"restaurant-management-backend/cmd/user/domain/types"
 )
@@ -17,6 +18,10 @@ func (this *UserDelete) Execute(id int) error {
 	userId, err := types.NewUserId(id)
 	if err != nil {
 		return err
+	}
+
+	if user, _ := this.repository.GetById(userId); user == nil {
+		return domain.NewUserNotFoundError(*userId)
 	}
 
 	return this.repository.Delete(userId)
